Python/PDF_Viewer_App_GUI/requirements.txt
Pillow
pdf2image